requirements.txt
# python==3.11.7
streamlit==1.44.1
pandasai==2.4.2
requests==1.5.3
python-dotenv==1.1.0
pandasai_llm==0.0.1